Combine PreviewSettings generated/custom code
